Move classes & some methods
